% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_album_plot.R
\name{album_plot}
\alias{album_plot}
\title{album_plot}
\usage{
album_plot(artist_data, ridge_variable, factor_variable)
}
\arguments{
\item{artist_data}{Data that is retrieved from spotifyr::get_artist_audio_features(). Each row of the 
data is a specific song. This plot will only utilize the numeric columns of the dataset.}

\item{ridge_variable}{Numeric variable from the artist_data that will be turned into
the joy ridges for the joy plot.}

\item{factor_variable}{Factor variable from the artist_data that will determine the 
grouping for the joy ridges on the y-axis.}
}
\value{
Returns a formatted ggjoy plot.
}
\description{
album_plot is a function that will take inputs for the data and
axis values and return a joy plot from the package 'ggjoy'.
}
\examples{
album_plot(spotifyr::get_artist_audio_features('AJR'),'Valence','album_name')

}
