% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentout.R
\name{alignmentout}
\alias{alignmentout}
\title{alignmentout}
\usage{
alignmentout(infile = "")
}
\arguments{
\item{infile}{entering the Mplus output file name and path}
}
\value{
A list of text, CSV files and one Excel file with multiple tabs in the specific folder.
}
\description{
Generating table(s) with estimates, alignment values and R-square of Thresholds and Loadings
}
\details{
Split a Mplus output into multiple parts.
Produce table(s) with estimates, alignment values and R-square of Thresholds and Loadings.
All would be stored in folder 'Output_current date' at the working directory.
}
\author{
Hai Nguyen \email{hnguye72@uic.edu}, Tianxiu Wang, Ariel Aloe, Rachel Gordon
}
