% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_lines}
\alias{cell_lines}
\title{List of metadata table and scaled PCs matrix}
\format{:
  meta_data: data.table of 9478 rows with information about dataset and cell_type
  scaled_pcs: data.table of 9478 rows (cells) and 20 columns (PCs)}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets}
}
\usage{
cell_lines
}
\description{
List of metadata table and scaled PCs matrix
}
\keyword{datasets}
