% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_lines_small}
\alias{cell_lines_small}
\title{Same as cell_lines but smaller (300 cells).}
\format{
An object of class \code{list} of length 2.
}
\source{
\url{support.10xgenomics.com/single-cell-gene-expression/datasets}
}
\usage{
cell_lines_small
}
\description{
Same as cell_lines but smaller (300 cells).
}
\keyword{datasets}
