% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_lines_small_sce}
\alias{cell_lines_small_sce}
\title{Same as cell_lines_small but as SingleCellExperiment object. 
Expression matrices filled in with dummy values.}
\format{An object of class \code{SingleCellExperiment} with 20 rows and 300 columns.}
\source{
\url{support.10xgenomics.com/single-cell-gene-expression/datasets}
}
\usage{
cell_lines_small_sce
}
\description{
Same as cell_lines_small but as SingleCellExperiment object. 
Expression matrices filled in with dummy values.
}
\keyword{datasets}
