% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_lines_small_seurat_v2}
\alias{cell_lines_small_seurat_v2}
\title{Same as cell_lines_small but as Seurat Version 2 object. 
Expression matrices filled in with dummy values.}
\format{An object of class \code{seurat} of length 1.}
\source{
\url{support.10xgenomics.com/single-cell-gene-expression/datasets}
}
\usage{
cell_lines_small_seurat_v2
}
\description{
Same as cell_lines_small but as Seurat Version 2 object. 
Expression matrices filled in with dummy values.
}
\keyword{datasets}
