% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celldetect.R
\name{cellsizedetect}
\alias{cellsizedetect}
\title{cellsizedetect}
\usage{
cellsizedetect(infile = "", n.detect = "0", silent = TRUE)
}
\arguments{
\item{infile}{entering the Mplus output file name and path}

\item{n.detect}{entering the number of cell size to detect, the default is zero}

\item{silent}{default is TRUE; if specifying FALSE, crosstabs are printed out}
}
\value{
A list of which crosstab(s) satisfied the input condition.
}
\description{
Check the 2 x 2 crosstabs of items having the cell size and below specified by the user
}
\details{
'cellsizedetect' requires Mplus output having CROSSTABS FOR CATEGORICAL VARIABLES part
}
\author{
Hai Nguyen \email{hnguye72@uic.edu}, Tianxiu Wang, Ariel Aloe, Rachel Gordon
}
