% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irtconvert.R
\name{convert2irt}
\alias{convert2irt}
\title{convert2irt}
\usage{
convert2irt(directory = NULL)
}
\arguments{
\item{directory}{entering the directory folder name to store the output}
}
\value{
A list of CSV files (with difficulty and discrimination of file name) in the specific folder.
}
\description{
Convert IFA estimates (threshold and loading) to IRT estimates (difficulty and discrimination)
}
\details{
convert2irt requires the user to firstly run the 'alignmentout' to obtain the threshold and loading parameters
}
\author{
Hai Nguyen \email{hnguye72@uic.edu}, Tianxiu Wang, Ariel Aloe, Rachel Gordon
}
