% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpcplot.R
\name{cpc}
\alias{cpc}
\title{cpc}
\usage{
cpc(selected.item = "", selected.group = "")
}
\arguments{
\item{selected.item}{selecting an item}

\item{selected.group}{selecting a group or multiple groups}
}
\value{
A graph including the category and cumulative probability curves stored in specific folder
}
\description{
Plotting the Category and Cumulative Probability Curve
}
\details{
cpc() requires the user to run 'alignmentout()' and 'convert2irt()' at first to obtain the difficulty and discrimination estimates
}
\author{
Hai Nguyen \email{hnguye72@uic.edu}, Ariel Aloe, Tianxiu Wang, Rachel Gordon
}
