% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpcvaluegrid.R
\name{cpcCPC.csvexport}
\alias{cpcCPC.csvexport}
\title{cpcCPC.csvexport}
\usage{
cpcCPC.csvexport(selected.item = "", selected.group = "", directory = NULL)
}
\arguments{
\item{selected.item}{selecting an item}

\item{selected.group}{selecting a group}

\item{directory}{entering the directory folder name to store the output}
}
\value{
A CSV file which has value grid of theta, category probability (cpc) and cumulative probability (CPC) values of a selected item's group
}
\description{
Producing the value grid to plot category and cumulative probability curve
}
\details{
cpcCPC.csvexport requires the user to run 'alignmentout()' and 'convert2irt()' at first to obtain the difficulty and discrimination estimates
}
\author{
Hai Nguyen \email{hnguye72@uic.edu}, Ariel Aloe, Tianxiu Wang, Rachel Gordon
}
