% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpcplot.R
\name{cpcCPC}
\alias{cpcCPC}
\title{cpcCPC}
\usage{
cpcCPC(selected.item = "", selected.group = "", directory = NULL)
}
\arguments{
\item{selected.item}{selecting an item}

\item{selected.group}{selecting a group or multiple groups}

\item{directory}{entering the directory folder name to store the output}
}
\value{
A graph including the category and cumulative probability curves stored in specific folder
}
\description{
Plotting the Category and Cumulative Probability Curve
}
\details{
cpcCPC() requires the user to run 'alignmentout()' and 'convert2irt()' at first to obtain the difficulty and discrimination estimates
}
\author{
Hai Nguyen \email{hnguye72@uic.edu}, Ariel Aloe, Tianxiu Wang, Rachel Gordon
}
