% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctrl.sparse}
\alias{ctrl.sparse}
\title{Sparse gene expression matrix of unstimulated PBMCs.}
\format{A sparse matrix (dgCMatrix) with 13019 cells (columns) and 35635 genes (rows):}
\source{
\url{https://www.nature.com/articles/nbt.4042}
}
\usage{
ctrl.sparse
}
\description{
Sparse gene expression matrix of unstimulated PBMCs.
}
\keyword{datasets}
