% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmony-package.R
\docType{package}
\name{harmony}
\alias{-package}
\alias{harmony}
\title{Harmony: fast, accurate, and robust single cell integration.}
\description{
Algorithm for single cell integration.
}
\section{Usage}{



?RunHarmony to run Harmony on cell embeddings matrix, Seurat or
SingleCellExperiment objects.
}

\section{Useful links}{


\enumerate{
\item Report bugs at \url{https://github.com/immunogenomics/harmony/issues}
\item Read the manuscript
\doi{10.64898/2026.03.16.711825}
}
}

