% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmony_option.R
\name{harmony_options}
\alias{harmony_options}
\title{Set advanced options for HarmonyMatrix}
\usage{
harmony_options(
  lambda = c(0.1, 10),
  tau = 0,
  block.size = 0.05,
  max.iter.cluster = 20,
  epsilon.cluster = 1e-05,
  epsilon.harmony = NULL
)
}
\arguments{
\item{lambda}{Default lambda = c(0.1, 10). In the new release of Harmony,
the appropriate value of lambda is dynamically estimated during harmony
iterations. And parameter lambda set the allowed range for lambda
estimation. e.g. lambda = c(0.1, 10) means that lambda can only vary
between 0.1 and 10.}

\item{tau}{Protection against overclustering small datasets with
large ones. `tau` is the expected number of cells per cluster.}

\item{block.size}{What proportion of cells to update during clustering. 
Between 0 to 1, default 0.05. Larger values may be faster but less
accurate.}

\item{max.iter.cluster}{Maximum number of rounds to run clustering
at each round of Harmony.}

\item{epsilon.cluster}{Convergence tolerance for clustering round
of Harmony. Set to -Inf to never stop early.}

\item{epsilon.harmony}{Convergence tolerance for Harmony. Set to -Inf to
never stop early. When `epsilon.harmony` is set to not NULL, then
user-supplied values of `early_stop` is ignored.}
}
\value{
Return a list for `.options` argument of `HarmonyMatrix`
}
\description{
Set advanced options for HarmonyMatrix
}
\examples{
## If want to set lambda to be fixed to 1, do
HarmonyMatrix(data_meta, meta_data, vars_use,
              .options = harmony_options(lambda = c(1, 1)))
}
