% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentsupportfnc.R
\name{invariancesplit}
\alias{invariancesplit}
\title{invariancesplit}
\usage{
invariancesplit(inputfile, filepath)
}
\arguments{
\item{inputfile}{usually a file of "Invariant_Noninvariant.txt" produced by a inside function from alignmentout}

\item{filepath}{a string of directory path to store the output files}
}
\value{
Two files of ThresholdInvariance.txt and LoadingsInvariance.txt in a specific folder
}
\description{
Split Invariance-Noninvariance result part into two sections of Intercepts/Thresholds and Loadings
}
\details{
Split Invariance-Noninvariance result part into two sections of Intercepts/Thresholds and Loadings and store each at the specified folder with names of ThresholdInvariance.txt and LoadingsInvariance.txt.
}
\author{
Hai Nguyen \email{hnguye72@uic.edu}
}
