% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ircvaluegrid.R
\name{irc.csvexport}
\alias{irc.csvexport}
\title{irc.csvexport}
\usage{
irc.csvexport()
}
\value{
A CSV file which has value grid of theta, category probability (cpc) and cumulative probability (CPC) values of a selected item's group
}
\description{
Producing the value grid to plot category and cumulative probability curve
}
\details{
irc.csvexport requires the user to run 'alignmentout()' and 'convert2irt()' at first to obtain the difficulty and discrimination estimates
}
\author{
Hai Nguyen \email{hnguye72@uic.edu}, Ariel Aloe, Tianxiu Wang, Rachel Gordon
}
