% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ircplot.R
\name{irc}
\alias{irc}
\title{irc}
\usage{
irc()
}
\value{
A graph including the category and cumulative probability curves in the Plots - R pane layout
}
\description{
Plotting the Category and Cumulative Probability Curve
}
\details{
irc() requires the user to run 'alignmentout()' and 'convert2irt()' at first to obtain the difficulty and discrimination estimates
}
\author{
Hai Nguyen \email{hnguye72@uic.edu}, Ariel Aloe, Tianxiu Wang, Rachel Gordon
}
