% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentsupportfnc.R
\name{latentsplit}
\alias{latentsplit}
\title{latentsplit}
\usage{
latentsplit(
  filepath = paste0("Output", "_", Sys.Date()),
  inputfile = "ext4_model results.txt"
)
}
\arguments{
\item{filepath}{usually a specific folder "Output_Date's today" stored the}

\item{inputfile}{usually a file of "ext4_model results.txt" produced by mplussplit function}
}
\value{
One or multiple latent class files depended on how many groups we have
}
\description{
Split Mplus latent classes into  separated parts and save each to a file
}
\details{
Splitting from the Mplus model result the latent classes into parts and store each at the specified folder with names of Latent Class text files.
}
\author{
Hai Nguyen \email{hnguye72@uic.edu}
}
