% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{moe_ridge_get_betas}
\alias{moe_ridge_get_betas}
\title{Get beta Utility}
\usage{
moe_ridge_get_betas(harmonyObj)
}
\arguments{
\item{harmonyObj}{Trained harmony object. Get this by running 
RunHarmony function with return_object=TRUE.}
}
\value{
Returns nothing, modifies object in place.
}
\description{
Utility function to get ridge regression coefficients from trained
Harmony object
}
