% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentsupportfnc.R
\name{mplussplit}
\alias{mplussplit}
\title{mplussplit}
\usage{
mplussplit(outpath = outpath, inputfile = inputfile)
}
\arguments{
\item{outpath}{usually set up as in the folder name Output_{date of running the program}}

\item{inputfile}{string of file path and Mplus output filename. It should be separated by / or \\}
}
\value{
Six output text files from an original Mplus output
}
\description{
Split a Mplus output into  multiple parts.
}
\details{
In order for the program to work, a Mplus output must have these parts:
1. INPUT INSTRUCTIONS
2. SUMMARY OF ANALYSIS
3. MODEL FIT INFORMATION
4. MODEL RESULTS
5. ALIGNMENT OUTPUT
6. SAVEDATA INFORMATION
Six files, including ext1_input instructions.txt, ext2_summary of analysis.txt, ext3_model fit information.txt, ext4_model results.txt, ext5_alignment output.txt, and ext6_savedata information.txt respectively will be generated and stored in the folder Output_{date of running the program} at the current working directory.
}
\author{
Hai Nguyen \email{hnguye72@uic.edu}
}
