% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentsupportfnc.R
\name{paraextract}
\alias{paraextract}
\title{paraextract}
\usage{
paraextract(inputfile, begphrase, endphrase, outputfile)
}
\arguments{
\item{inputfile}{a string of file path and Mplus output file name. It should be separated by / or \\}

\item{begphrase}{a phrase/words the wanted paragraph started with}

\item{endphrase}{a phrase/words the wanted paragraph ended at}

\item{outputfile}{a string of file path and output file name to store in}
}
\value{
A specific paragraph from origin text file
}
\description{
Extract a specific paragraph from a text file
}
\details{
Extract a wanted paragraph starting with the beginning phrase and ending at the ending phrase. This based on the algorithm from stackoverflow of
\url{https://stackoverflow.com/questions/28676208/how-to-extract-specific-paragraphs-from-a-text-file-in-r}. But, this one was modified to write the paragraph to text file for reference.
}
\author{
Hai Nguyen \email{hnguye72@uic.edu}
}
