% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc.small}
\alias{pbmc.small}
\title{List of expression matrix, metadata table, and PCA embeddings for 500 stimulated and 500 unstimulated PBMCs.}
\format{List of 3 objects: 
  1) exprs_norm_vargenes: dgCMatrix of 1636 genes by 1000 cells. 
  2) meta_data: data.frame
  3) pca_embeddings: matrix of 1000 cells by 20 PCs}
\source{
\url{https://www.nature.com/articles/nbt.4042}
}
\usage{
pbmc.small
}
\description{
List of expression matrix, metadata table, and PCA embeddings for 500 stimulated and 500 unstimulated PBMCs.
}
\keyword{datasets}
