% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc.small.seurat}
\alias{pbmc.small.seurat}
\title{Seurat object with the same data as pbmc.small.}
\format{Seurat v2.3.4 object.}
\source{
\url{https://www.nature.com/articles/nbt.4042}
}
\usage{
pbmc.small.seurat
}
\description{
Seurat object with the same data as pbmc.small.
}
\keyword{datasets}
