% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc.stim}
\alias{pbmc.stim}
\title{Gene expression data of stimulated PBMC from Kang et al. 2017. This
contains a sample of 1000 cells from that condition and is used for
the Seurat Vignette.}
\format{
An object of class \code{dgCMatrix} with 9015 rows and 1000 columns.
}
\source{
\doi{10.1038/nbt.4042}
}
\usage{
pbmc.stim
}
\description{
Gene expression data of stimulated PBMC from Kang et al. 2017. This
contains a sample of 1000 cells from that condition and is used for
the Seurat Vignette.
}
\keyword{datasets}
