% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pca_plot.R
\name{pca_plot}
\alias{pca_plot}
\title{pca_plot}
\usage{
pca_plot(data, album_filter)
}
\arguments{
\item{data}{A filtered dataframe from the spotifyr function 
get_artist_audio_features.}

\item{album_filter}{A list of albums to perform the PCA on and plot.}
}
\value{
Returns a formatted ggplot.
}
\description{
A plotting function that receives a data frame and runs
PCA on it to produce a formatted ggplot.
}
\examples{
pca_plot(spotifyr::get_artist_audio_features('Morgan Wallen'), c('Dangerous: The Double Album', 'If I Know Me'))

}
