% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pca_table.R
\name{pca_table}
\alias{pca_table}
\title{pca_table}
\usage{
pca_table(data, album_filter)
}
\arguments{
\item{data}{A data frame from the spotifyr function '
get_artist_audio_features.' PCA will be performed on this data within the 
function.}

\item{album_filter}{A character vector of album names that the user selects
to display on the plot created by the function 'pca_plot'.}
}
\value{
TThe return value of this function is a sorted data frame. The data 
frame contains the values of all 9 of the principal components, as well as 
the album name and song name for each value. The row names are the original 
numbers from the input data frame.
}
\description{
A function to create a table with the values that are plotted 
by the pca_plot module. This function will work inside of a renderTable 
statement to display the data from the PCA that is done in the plot and
help the user identify outlier songs.
}
\examples{
pca_table(spotifyr::get_artist_audio_features('Adele'),c('25','19'))

}
