% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_song_plot.R
\name{song_plot}
\alias{song_plot}
\title{song_plot}
\usage{
song_plot(artist_data, plot_xaxis, plot_yaxis, color)
}
\arguments{
\item{artist_data}{Data that is retrieved from spotifyr::get_artist_audio_features(). Each row of the 
data is a specific song. This plot will only utilize the numeric columns of the dataset.}

\item{plot_xaxis, plot_yaxis}{Specify the x-axis and y-axis of the plot. Must be a numeric variable from artist_data.}

\item{color}{Sets the color of the points in the plot. Should be a factor or character variable from 
the data. Numeric colors will give the plot a continuous color scale.}
}
\value{
Returns a formatted ggplot object.
}
\description{
song_plot is a function that will take inputs for data and 
axis variables and return a ggplot.
}
\examples{
 
Sys.setenv(SPOTIFY_CLIENT_ID = '3e2ffc9154ba4bdf885282b6afab8e1a')
Sys.setenv(SPOTIFY_CLIENT_SECRET = '5325176284b34d2a9ab13ececac8a417')
data <- spotifyr::get_artist_audio_features('The Head and the Heart')

ggsong <- song_plot(data,'valence','danceability', 'album_name')
ggsong

}
