% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stim.sparse}
\alias{stim.sparse}
\title{Sparse gene expression matrix of stimulated PBMCs.}
\format{A sparse matrix (dgCMatrix) with 12875 cells (columns) and 35635 genes (rows):}
\source{
\url{https://www.nature.com/articles/nbt.4042}
}
\usage{
stim.sparse
}
\description{
Sparse gene expression matrix of stimulated PBMCs.
}
\keyword{datasets}
