% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celldetect.R
\name{cellsizedetect}
\alias{cellsizedetect}
\title{cellsizedetect}
\usage{
cellsizedetect(infile = "")
}
\arguments{
\item{infile}{entering the Mplus output file name and path}
}
\value{
Crosstabs between items having the specified cell size and below
}
\description{
Check the 2 x 2 crosstabs of items having the cell size and below specified by the user
}
\details{
'cellsizedetect' requires Mplus output having CROSSTABS FOR CATEGORICAL VARIABLES part
}
\author{
Hai Nguyen \email{hnguye72@uic.edu}, Tianxiu Wang, Ariel Aloe, Rachel Gordon
}
