% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmony-package.r
\docType{package}
\name{harmony}
\alias{harmony}
\title{Harmony: fast, accurate, and robust single cell integration.}
\description{
Algorithm for single cell integration.
}
\section{Usage}{
 

\enumerate{
\item ?HarmonyMatrix to run Harmony on gene expression or PCA 
embeddings matrix.
\item ?RunHarmony to run Harmony on Seurat objects.
}
}

\section{Useful links}{


\enumerate{
\item Report bugs at \url{https://github.com/immunogenomics/harmony/issues}
\item Read the manuscript
\href{https://www.nature.com/articles/s41592-019-0619-0}{online}.
}
}

