% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmony-package.R
\docType{package}
\name{harmony}
\alias{harmony-package}
\alias{harmony}
\title{Harmony: fast, accurate, and robust single cell integration.}
\description{
Algorithm for single cell integration.
}
\section{Usage}{



?RunHarmony to run Harmony on cell embeddings matrix, Seurat or
SingleCellExperiment objects.
}

\section{Useful links}{


\enumerate{
\item Report bugs at \url{https://github.com/immunogenomics/harmony/issues}
\item Read the manuscript
\doi{10.64898/2026.03.16.711825}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/immunogenomics/harmony}
  \item \url{https://pati-ni.github.io/harmony/}
}

}
\author{
\strong{Maintainer}: Ilya Korsunsky \email{ilya.korsunsky@gmail.com} (\href{https://orcid.org/0000-0003-4848-3948}{ORCID})

Authors:
\itemize{
  \item Nikolaos Patikas \email{nik.patik@gmail.com} (\href{https://orcid.org/0000-0002-3978-0134}{ORCID}) [contributor]
  \item Hongcheng Yao \email{hongchengyaonk@gmail.com} (\href{https://orcid.org/0000-0002-0743-4835}{ORCID}) [contributor]
  \item Martin Hemberg \email{mhemberg@bwh.harvard.edu} (\href{https://orcid.org/0000-0001-8895-5239}{ORCID})
  \item Nghia Millard \email{nmillard@g.harvard.edu} (\href{https://orcid.org/0000-0002-0518-7674}{ORCID})
  \item Jean Fan \email{jeanfan@fas.harvard.edu} (\href{https://orcid.org/0000-0002-0212-5451}{ORCID}) [contributor]
  \item Kamil Slowikowski \email{kslowikowski@gmail.com} (\href{https://orcid.org/0000-0002-2843-6370}{ORCID}) [contributor]
  \item Soumya Raychaudhuri (\href{https://orcid.org/0000-0002-1901-8265}{ORCID})
}

Other contributors:
\itemize{
  \item Miles Smith [contributor]
}

}
\keyword{internal}
