% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarmonyMatrix.R
\name{HarmonyMatrix}
\alias{HarmonyMatrix}
\title{HarmonyMatrix}
\usage{
HarmonyMatrix(pc_mat, meta_data, vars_use, theta = NULL, lambda = NULL,
  sigma = 0.1, alpha = 0.1, nclust = 100, tau = 0,
  block.size = 0.05, max.iter.harmony = 10, max.iter.cluster = 200,
  epsilon.cluster = 1e-05, epsilon.harmony = 1e-04,
  burn.in.time = 10, plot_convergence = FALSE, return_object = FALSE)
}
\arguments{
\item{alpha}{strength of dirichlet prior for DKL penalty}

\item{tau}{desired cells/cluster}

\item{plot_convergence}{window size for kmeans convergenc}
}
\description{

}
