% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ircplot.R
\name{irc}
\alias{irc}
\title{irc}
\usage{
irc()
}
\value{
A list of Item Response (Characteristic) Curve Plot in R environment
}
\description{
Generating Cumulative and Category Item Response (Characteristic) Curve Plots
}
\details{
irc requires the user to firstly run the 'alignmentout' and 'convert2irt' to obtain the difficulty and discrimination parameters
}
\author{
Hai Nguyen \email{hnguye72@uic.edu}, Ariel Aloe, Tianxiu Wang, Rachel Gordon
}
