% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{RunHarmony}
\alias{RunHarmony}
\title{Harmony wrapper for Seurat}
\usage{
RunHarmony(object, group.by.vars, dims.use, theta = NULL,
  lambda = NULL, sigma = 0.1, nclust = 100, tau = 0,
  block.size = 0.05, max.iter.harmony = 10, max.iter.cluster = 20,
  epsilon.cluster = 1e-05, epsilon.harmony = 1e-04,
  plot_convergence = FALSE, verbose = TRUE, reference_values = NULL)
}
\arguments{
\item{object}{Seurat object. Must have PCA computed.}

\item{group.by.vars}{Which variable(s) to remove (character vector).}

\item{theta}{Diversity clustering penalty parameter. Specify for each variable in group.by.vars. Default theta=2. theta=0 does not encourage any diversity. Larger values of theta result in more diverse clusters.}

\item{lambda}{Ridge regression penalty parameter. Specify for each variable in group.by.vars. Default lambda=1. Lambda must be strictly positive. Smaller values result in more aggressive correction.}

\item{sigma}{Width of soft kmeans clusters. Default sigma=0.1. Sigma scales the distance from a cell to cluster centroids. Larger values of sigma result in cells assigned to more clusters. Smaller values of sigma make soft kmeans cluster approach hard clustering.}

\item{nclust}{Number of clusters in Harmony model. nclust=1 equivalent to simple linear regression.}

\item{tau}{Protection against overclustering small datasets with large ones. tau is the expected number of cells per cluster.}

\item{block.size}{What proportion of cells to update during clustering. Between 0 to 1, default 0.05. Larger values may be faster but less accurate.}

\item{max.iter.harmony}{Maximum number of rounds to run Harmony. One round of Harmony involves one clustering and one correction step.}

\item{max.iter.cluster}{Maximum number of rounds to run clustering at each round of Harmony.}

\item{epsilon.cluster}{Convergence tolerance for clustering round of Harmony. Set to -Inf to never stop early.}

\item{epsilon.harmony}{Convergence tolerance for Harmony. Set to -Inf to never stop early.}

\item{plot_convergence}{Whether to print the convergence plot of the clustering objective function. TRUE to plot, FALSE to suppress. This can be useful for debugging.}

\item{verbose}{Whether to print progress messages. TRUE to print, FALSE to suppress.}

\item{reference_values}{(Advanced Usage) Defines reference dataset(s). Cells that have batch variables values matching reference_values will not be moved.}
}
\value{
Seurat object. Harmony dimensions placed into object@dr$harmony. For downstream Seurat analyses, use reduction.use='harmony' and reduction.type='harmony'.
}
\description{
Use this to run the Harmony algorithm on a Seurat object. See also HarmonyMatrix to run directly on PCA matrix.
}
